fix: Move codes in densitymatrix_simulation_helpers.py to densitymatrix_simulation.py
